Implement toplevel isna and move tests
